# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ArrControl API
  version: 0.1.0
  license: { name: MIT }
servers: [{ url: /api/v1 }]
security: [{ cookieAuth: [] }]
paths:
...
  /instances/{instanceId}/credentials:
    get:
      operationId: listInstanceCredentials
      description: Lists only write-only credential metadata for an instance visible through `instances.read`. Plaintext and encryption material are never returned.
      parameters: [{ $ref: '#/components/parameters/InstanceId' }]
      responses:
        '200':
          description: Configured credential purposes
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/CredentialMetadata' } } } }
        '400': { $ref: '#/components/responses/CredentialRequestInvalid' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/InstanceNotFound' }
  /instances/{instanceId}/credentials/{purpose}:
    parameters:
      - { $ref: '#/components/parameters/InstanceId' }
      - { $ref: '#/components/parameters/CredentialPurpose' }
    put:
      operationId: putInstanceCredential
      description: Creates or replaces a write-only credential using the target instance's effective `instances.manage` scope. Requires CSRF validation; the request is limited to 8 KiB.
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/WriteCredential' } } }
      responses:
        '200':
          description: Existing credential replaced
          content: { application/json: { schema: { $ref: '#/components/schemas/CredentialMetadata' } } }
        '201':
          description: Credential created
          headers: { Location: { schema: { type: string, format: uri-reference } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/CredentialMetadata' } } }
        '400': { $ref: '#/components/responses/CredentialRequestInvalid' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AuthorizationMutationForbidden' }
        '404': { $ref: '#/components/responses/InstanceNotFound' }
        '503': { $ref: '#/components/responses/CredentialEncryptionUnavailable' }
    delete:
      operationId: deleteInstanceCredential
      description: Idempotently deletes one credential using the target instance's effective `instances.manage` scope. Requires CSRF validation.
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: Credential deleted or already absent }
        '400': { $ref: '#/components/responses/CredentialRequestInvalid' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AuthorizationMutationForbidden' }
        '404': { $ref: '#/components/responses/InstanceNotFound' }
...
  /queue:
    get:
      operationId: listQueue
      parameters: [{ name: instanceId, in: query, schema: { type: array, maxItems: 100, items: { type: string, format: uuid } } }]
      responses:
        '200': { description: Aggregated current queue with history correlation, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AggregatedQueueItem' } } } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
  /history:
    get:
      operationId: listHistory
      parameters:
        - { name: instanceId, in: query, schema: { type: array, maxItems: 100, items: { type: string, format: uuid } } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 200 } }
      responses:
        '200': { description: Aggregated recent history with current-queue correlation, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AggregatedHistoryItem' } } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: __Host-arrcontrol_session, description: Short-lived opaque access token in a Secure, host-only, Path=/, HttpOnly, SameSite cookie. }
    refreshCookie: { type: apiKey, in: cookie, name: __Host-arrcontrol_refresh, description: Rotating opaque refresh token in a Secure, host-only, Path=/, HttpOnly, SameSite cookie. }
  parameters:
    InstanceId: { name: instanceId, in: path, required: true, schema: { type: string, format: uuid } }
...
    CredentialPurpose: { name: purpose, in: path, required: true, schema: { $ref: '#/components/schemas/CredentialPurpose' } }
...
    CsrfToken: { name: X-CSRF-Token, in: header, required: true, description: Must exactly match the __Host-arrcontrol_csrf cookie., schema: { type: string, minLength: 43, maxLength: 43 } }
  responses:
    Problem: { description: RFC 9457 problem details, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    AuthenticationRequired: { description: 'Authentication is required (`authentication_required`).', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    AccessDenied: { description: 'The effective authorization does not grant access (`access_denied`).', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    AuthorizationMutationForbidden: { description: 'Access is denied (`access_denied`) or CSRF validation failed (`csrf_validation_failed`).', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    CredentialRequestInvalid: { description: 'The credential target, purpose, or secret is invalid (`credential_target_invalid`, `credential_purpose_invalid`, or `credential_secret_invalid`).', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    InstanceNotFound: { description: 'The instance does not exist or is outside the caller scope (`instance_not_found`).', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    CredentialEncryptionUnavailable: { description: 'No valid active credential encryption key is configured (`credential_encryption_unavailable`).', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
...
    CredentialPurpose: { type: string, enum: [api-key,username,password] }
    WriteCredential:
      type: object
      additionalProperties: false
      required: [secret]
      properties:
        secret: { type: string, minLength: 1, maxLength: 4096, writeOnly: true, description: Maximum is 4,096 UTF-8 bytes; the character limit is only an OpenAPI upper bound. }
    CredentialMetadata:
      type: object
      additionalProperties: false
      required: [purpose,configured,updatedAt]
      properties:
        purpose: { $ref: '#/components/schemas/CredentialPurpose' }
        configured: { const: true }
        updatedAt: { type: string, format: date-time }
...
    AggregatedQueueItem:
      type: object
      additionalProperties: false
      required: [instanceId,instanceName,providerKind,providerKey,mediaProviderKey,title,status,trackedStatus,trackedState,protocol,sizeBytes,remainingBytes,addedAt,estimatedCompletionAt,downloadClient,indexer,correlatedHistoryCount,latestHistoryEvent,importFailure,observedAt,stale]
      properties:
        instanceId: { type: string, format: uuid }
        instanceName: { type: string }
        providerKind: { type: string, enum: [sonarr,radarr,lidarr,readarr,whisparr,prowlarr,sabnzbd,nzbget,qbittorrent,transmission,deluge] }
        providerKey: { type: string }
        mediaProviderKey: { type: [string, 'null'] }
        title: { type: string }
        status: { type: string }
        trackedStatus: { type: string }
        trackedState: { type: string }
        protocol: { type: [string, 'null'] }
        sizeBytes: { type: integer, format: int64, minimum: 0 }
        remainingBytes: { type: integer, format: int64, minimum: 0 }
        addedAt: { type: [string, 'null'], format: date-time }
        estimatedCompletionAt: { type: [string, 'null'], format: date-time }
        downloadClient: { type: [string, 'null'] }
        indexer: { type: [string, 'null'] }
        correlatedHistoryCount: { type: integer, minimum: 0 }
        latestHistoryEvent: { type: [string, 'null'] }
        importFailure: { oneOf: [{ $ref: '#/components/schemas/ImportFailureClassification' }, { type: 'null' }] }
        observedAt: { type: string, format: date-time }
        stale: { type: boolean }
    AggregatedHistoryItem:
      type: object
      additionalProperties: false
      required: [instanceId,instanceName,providerKind,providerKey,mediaProviderKey,title,eventType,eventAt,queueCorrelated,importFailure,observedAt,stale]
      properties:
        instanceId: { type: string, format: uuid }
        instanceName: { type: string }
        providerKind: { type: string, enum: [sonarr,radarr,lidarr,readarr,whisparr,prowlarr,sabnzbd,nzbget,qbittorrent,transmission,deluge] }
        providerKey: { type: string }
        mediaProviderKey: { type: [string, 'null'] }
        title: { type: string }
        eventType: { type: string }
        eventAt: { type: string, format: date-time }
        queueCorrelated: { type: boolean }
        importFailure: { oneOf: [{ $ref: '#/components/schemas/ImportFailureClassification' }, { type: 'null' }] }
        observedAt: { type: string, format: date-time }
        stale: { type: boolean }
    ImportFailureClassification:
      type: object
      additionalProperties: false
      required: [category,severity,ruleId,guidanceKey,retrySupported]
      properties:
        category: { type: string, enum: [download_client_unavailable,import_blocked,import_failed,import_retry_pending,upstream_error,import_pending,download_failed,download_ignored] }
        severity: { type: string, enum: [warning,error] }
        ruleId: { type: string }
        guidanceKey: { type: string, description: Localization key; the API does not synthesize a diagnosis from free text. }
        retrySupported: { const: false, description: Remains false until a safe provider retry contract is implemented and tested. }
...
    Problem: { type: object, additionalProperties: false, required: [type,title,status,code,traceId], properties: { type: { type: string, format: uri-reference }, title: { type: string }, status: { type: integer }, detail: { type: string }, traceId: { type: string }, code: { type: string } } }